+++ src/FsCodec.NewtonsoftJson/Serdes.fs
@@ -69,6 +69,10 @@ type Serdes(options : JsonSerializerSettings) =
         use jsonReader = Utf8BytesEncoder.makeJsonReader ms
         serializer.Deserialize<'T>(jsonReader)
 
+    /// Deserializes value of given type from a JObject
+    member _.Deserialize<'T>(jo : Newtonsoft.Json.Linq.JObject) : 'T =
+        jo.ToObject(typeof<'T>, serializer) :?> 'T
+
     /// Serializes and writes given value to a stream.
     member _.SerializeToStream<'T>(value : 'T, utf8Stream : Stream) =
         // We're setting CloseOutput = false, because that's the default behavior in STJ