+++ src/FsCodec.SystemTextJson/Pickler.fs
@@ -2,31 +2,11 @@
 
 open System.Text.Json
 
-[<AutoOpen>]
-module private Prelude =
-    let memoize (f: 'T -> 'S): 'T -> 'S =
-        let cache = new System.Collections.Concurrent.ConcurrentDictionary<'T, 'S>()
-        fun t -> cache.GetOrAdd(t, f)
-
 [<AbstractClass>]
 type JsonPickler<'T>() =
     inherit Serialization.JsonConverter<'T>()
 
-    static let isMatchingType =
-        let rec isMatching = function
-            | [] -> false
-            | t :: _ when t = typeof<'T> -> true
-            | t :: tl ->
-                let tail =
-                    [ match t.BaseType with null -> () | bt -> yield bt
-                      yield! t.GetInterfaces()
-                      yield! tl ]
-                isMatching tail
-        memoize (fun t -> isMatching [t])
-
     abstract Read: reader: byref<Utf8JsonReader> * options: JsonSerializerOptions -> 'T
-
-    override _.CanConvert t = isMatchingType t
 
     override x.Read(reader, _typeToConvert, options) =
         x.Read(&reader, options)