+++ src/FsCodec.SystemTextJson/UnionOrTypeSafeEnumConverterFactory.fs
@@ -5,12 +5,13 @@ open System.Text.Json.Serialization
 type UnionOrTypeSafeEnumConverterFactory(typeSafeEnum, union) =
     inherit JsonConverterFactory()
 
-    static let hasConverterAttribute: System.Type -> bool = memoize _.IsDefined(typeof<JsonConverterAttribute>, true)
+    static let cache = System.Collections.Concurrent.ConcurrentDictionary<System.Type, bool>()
+    static let typeHasConverterAttribute t: bool = cache.GetOrAdd(t, fun (t: System.Type) -> t.IsDefined(typeof<JsonConverterAttribute>, ``inherit`` = false))
 
     override _.CanConvert t =
         not (t.IsGenericType && let g = t.GetGenericTypeDefinition() in g = typedefof<option<_>> || g = typedefof<list<_>>)
         && FsCodec.Union.isUnion t
-        && not (hasConverterAttribute t)
+        && not (typeHasConverterAttribute t)
         && ((typeSafeEnum && union)
             || typeSafeEnum = FsCodec.Union.isNullary t)