+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The `Unreleased` section name is replaced by the expected version of next releas
 ### Added
  
 - `NewtonsoftJson.Options.CreateStringEnumConverter`: Small wrapper to smooth construction syntax [#109](https://github.com/jet/FsCodec/pull/109)
+- `NewtonsoftJson.Serdes.Deserialize<'T>(Newtonsoft.Json.Linq.JObject)`: One liner [#109](https://github.com/jet/FsCodec/pull/109)
 
 ### Changed