+++ CHANGELOG.md
@@ -10,8 +10,8 @@ The `Unreleased` section name is replaced by the expected version of next releas
 
 ### Added
  
-- `NewtonsoftJson.Options.CreateStringEnumConverter`: Small wrapper to smooth construction syntax [#109](https://github.com/jet/FsCodec/pull/109)
-- `NewtonsoftJson.Serdes.Deserialize<'T>(Newtonsoft.Json.Linq.JObject)`: One liner [#109](https://github.com/jet/FsCodec/pull/109)
+- `NewtonsoftJson.Options.CreateStringEnumConverter`: Small wrapper to smooth construction syntax [#111](https://github.com/jet/FsCodec/pull/111)
+- `NewtonsoftJson.Serdes.Deserialize<'T>(Newtonsoft.Json.Linq.JObject)`: One liner [#111](https://github.com/jet/FsCodec/pull/111)
 
 ### Changed